+++ src/Support/RuleEngine.php
@@ -8,6 +8,8 @@ use Sufyan\MigrationLinter\Rules\FloatColumnForMoney;
 use Sufyan\MigrationLinter\Rules\DropColumnWithoutBackup;
 use Sufyan\MigrationLinter\Rules\MissingIndexOnForeignKey;
 use Sufyan\MigrationLinter\Rules\SoftDeletesOnProduction;
+use Sufyan\MigrationLinter\Rules\RenamingColumnWithoutIndex;
+use Sufyan\MigrationLinter\Rules\ChangeColumnTypeOnLargeTable;
 use Sufyan\MigrationLinter\Rules\AddNonNullableColumnWithoutDefault;
 use Sufyan\MigrationLinter\Rules\AddUniqueConstraintOnNonEmptyColumn;
 
@@ -44,6 +46,8 @@ class RuleEngine
             'AddUniqueConstraintOnNonEmptyColumn' => AddUniqueConstraintOnNonEmptyColumn::class,
             'FloatColumnForMoney' => FloatColumnForMoney::class,
             'SoftDeletesOnProduction' => SoftDeletesOnProduction::class,
+            'RenamingColumnWithoutIndex' => RenamingColumnWithoutIndex::class,
+            'ChangeColumnTypeOnLargeTable' => ChangeColumnTypeOnLargeTable::class,
         ];
 
         // Load config-defined rules (allows app or custom packages)